Add 64 bits integer
We should take care : use the exact same integer type in C and fortran.

+++ c_fortran/c.c
@@ -1,9 +1,10 @@
-// file: c_function.c
-
+#include <inttypes.h>
 #include <stdio.h>
 
-void c_function(int* array, int n) {
-  int i;
+
+
+void c_function(int64_t* array, int64_t n) {
+  int64_t i;
   for(i = 0; i < n; i++) {
     printf("%d\n", array[i]);
   }
